bulk-backend/requirements.txt
fastapi==0.109.0
uvicorn==0.24.0
sqlalchemy==2.0.25
pymysql==1.1.0  
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
pydantic==2.6.4
python-multipart==0.0.9
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.3
redis==5.0.1
celery==5.3.6
uuid==1.30